fix(citizen/devtools): x86 buildability
Thanks to @GottfriedLeibniz for reporting!

+++ code/components/citizen-devtools/src/ResourceMonitor.cpp
@@ -86,14 +86,14 @@ namespace fx
 struct ResourceMonitorImpl : public ResourceMonitorImplBase
 {
 	ResourceMonitorImpl()
-		: tickIndex(std::make_shared<size_t>(0)), scriptFrameMetrics(tickIndex), gameFrameMetrics(tickIndex)
+		: tickIndex(std::make_shared<uint64_t>(0)), scriptFrameMetrics(tickIndex), gameFrameMetrics(tickIndex)
 	{
 		
 	}
 
 	virtual ~ResourceMonitorImpl() = default;
 
-	std::shared_ptr<size_t> tickIndex;
+	std::shared_ptr<uint64_t> tickIndex;
 
 	tbb::concurrent_unordered_map<std::string, std::optional<ResourceMetrics>> metrics;
 	TickMetrics<64> scriptFrameMetrics;
